refactor: change structure of main page
